[ALPS04239425] Camera: fix build error
Fix build error in kernel-4.14

MTK-Commit-Id: 464e9d58948f3f5d14740a45384022e0b2f893d4

CR-Id: ALPS04239425
Feature: Build System

+++ drivers/misc/mediatek/sensors-1.0/accelerometer/accel.c
@@ -267,7 +267,7 @@ static int acc_enable_and_batch(void)
 		/* start polling, if needed */
 		if (cxt->is_active_data == true &&
 		    cxt->acc_ctl.is_report_input_direct == false) {
-			int mdelay = cxt->delay_ns;
+			uint64_t mdelay = cxt->delay_ns;
 
 			do_div(mdelay, 1000000);
 			atomic_set(&cxt->delay, mdelay);
